move data files into directory
