+++ src/modules/english/data/englishVerbDictionary.ts
